requirements.txt
accelerate @ git+https://github.com/huggingface/accelerate.git@c0552c9012a9bae7f125e1df89cf9ee0b0d250fd
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiolimiter==1.1.0
aiosignal==1.3.1
alembic==1.14.0
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncio==3.4.3
attrs==24.2.0
Authlib==1.3.2
bandit==1.7.10
bcrypt==4.2.0
bitsandbytes==0.42.0
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cohere==5.11.3
contourpy==1.3.1
coverage==7.6.4
cryptography==43.0.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
databases==0.9.0
dataclasses-json==0.6.7
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
dparse==0.6.4
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.5
fastavro==1.9.7
filelock==3.12.4
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
identify==2.6.2
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.7
langchain-cohere==0.3.1
langchain-community==0.3.7
langchain-core==0.3.18
langchain-experimental==0.3.3
langchain-text-splitters==0.3.2
langsmith==0.1.143
loguru==0.7.2
macholib==1.16.3
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy==1.13.0
mypy-extensions==1.0.0
networkx==3.4.2
ninja==1.11.1.1
nodeenv==1.9.1
numpy==1.26.4
openai==1.54.4
orjson==3.10.11
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
passlib==1.7.4
pathspec==0.12.1
pbr==6.1.0
peft @ git+https://github.com/huggingface/peft.git@221965b7e140bdd47fbd04c131d56029e089902c
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
propcache==0.2.0
protobuf==5.28.3
psutil==6.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
quanto==0.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
safety==3.2.11
safety-schemas==0.0.9
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
setuptools==75.5.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sse-starlette==2.1.3
starlette==0.41.2
starlette-context==0.3.6
stevedore==5.3.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
tqdm==4.67.0
typer==0.13.0
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.27.1
voyageai==0.3.1
yarl==1.17.1
python-docx